Rework to order structuring, using clsOrderCollection to store a list of clsOrder, each clsOrder storing a list of clsOrderLines (when necessary).
Tests haven't all been reworked to work with this structure and may fail upon running.

+++ Testing3/tstOrder.cs
@@ -36,7 +36,7 @@ namespace Testing3
         public void TotalCostPropertyOK()
         {
             clsOrder NewOrder = new clsOrder();
-            float TestData = 399.99f;
+            double TestData = 399.99f;
             NewOrder.SetTotalCost(TestData);
             Assert.AreEqual(NewOrder.GetTotalCost(), TestData);
         }
